Initial Build of Project
This is the first build/prototype of a lyrics translation site. I
played around with the functionality and two options, one using jQuery
and buttons and the other(tabs.html) using jQuery UI. I’m not sure
which I prefer or will be better long term. The second round will
include styling elements in more depth.